+++ docs/Changelog.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## v2.3.2
+* Disable auto feed refresh by default and add setting to re-enable it
+* Update theme
+* Disable bugsnag completely when opting out of analytics
+
 ## v2.3.1
 * Hide all story panels if enabled
 * Prevent swipe to refresh if not at the very top